+++ changelog.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Fixed
 
 - Correct MSRV failure in CI build (`6354036`)
-- Validate `DragOptions::bounds` (``)
+- Validate `DragOptions::bounds` (`9f0ec17`)
+- Add CDP drag bound clamping tests (``)
 
 # [Released]